requirements.txt
pyusb==1.0.0b2
kafka-python==0.9.4
